add ls command and make some fixes
we need to use a different salt for the hash of the encrypted file
contents, and it's also helpful for the ls command to store the output
file name from when the file was encrypted, so the qlock file.

we also now check if the file already exists during encryption too

definitely some room to make things more DRY (Dont Repeat Yourself) but
it's okay for now.

+++ src/main.rs
@@ -17,13 +17,24 @@ use argon2::Algorithm::Argon2d;
 use argon2::Version::V0x13;
 use argon2::{Argon2, Params};
 
-use clap::Parser;
+use clap::{Parser, Subcommand};
 
 #[derive(Parser)]
 #[command(author, version, about, long_about = Some("things adn stufffffffff fffff fff f f f f f f ff f f ffffffff f "))]
 struct Qlock {
+    /// Commands
+    #[command(subcommand)]
+    command: Option<Commands>,
+
+    #[command(flatten)]
+    action: Option<ActionArgs>,
+}
+
+#[derive(Parser)]
+struct ActionArgs {
     /// File to encrypt or decrypt
     #[arg(value_parser = clap::value_parser!(PathBuf))]
+    #[arg(required = false)]
     file: PathBuf,
 
     /// Encrypt the input file
@@ -39,6 +50,12 @@ struct Qlock {
     output: Option<String>,
 }
 
+#[derive(Subcommand)]
+enum Commands {
+    /// list the hashes, input filename, and output filename, for each file you've already encrypted
+    Ls,
+}
+
 #[derive(Serialize, Deserialize, Debug)]
 struct EncryptedData {
     hash: Vec<u8>,
@@ -46,7 +63,9 @@ struct EncryptedData {
     nonce_a: Vec<u8>,
     nonce_b: Vec<u8>,
     salt: Vec<u8>,
+    hash_salt: Vec<u8>,
     filename: String,
+    output_filename: String,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -57,32 +76,42 @@ struct SavedData {
 fn main() -> std::io::Result<()> {
     let cli = Qlock::parse();
 
-    if cli.encrypt {
-        encrypt_cmd(&cli.file, cli.output);
-    } else if cli.decrypt {
-        decrypt_cmd(&cli.file, cli.output);
-    } else {
-        eprintln!("Please specify either --encrypt or --decrypt");
-        std::process::exit(1);
+    match cli.command {
+        Some(Commands::Ls) => {
+            list_metadata();
+            return Ok(());
+        }
+        None => {}
+    }
+
+    if let Some(action) = cli.action {
+        match (action.encrypt, action.decrypt) {
+            (true, false) => {
+                encrypt_cmd(&action.file, action.output);
+            }
+            (false, true) => {
+                decrypt_cmd(&action.file, action.output);
+            }
+            _ => {
+                eprintln!("Please specify either --encrypt or --decrypt");
+                std::process::exit(1);
+            }
+        }
     }
 
     Ok(())
 }
 
 fn encrypt_cmd(file_path: &PathBuf, output_path: Option<String>) {
-    // 1. prompt the user for a password
     let prompt_text = "Don't forget to backup your password!\n\nIf you forget your password, you will not be able to decrypt your files!\n\n(min 16 chars, mix of upper + lower case, at least 1 number or special character)\nCreate a new password: ";
     let password = rpassword::prompt_password(prompt_text).unwrap();
 
-    // 2. get the contents of the file at the provided file path
     let contents = fs::read(file_path).expect("oops, couldn't read the file");
 
-    // 3. generate a key
     let og_key = XChaCha20Poly1305::generate_key(&mut OsRng);
     let cipher_a = XChaCha20Poly1305::new(&og_key);
     let nonce_a = XChaCha20Poly1305::generate_nonce(&mut OsRng);
 
-    // 4. encrypt the contents of the file
     let ciphertext = cipher_a
         .encrypt(&nonce_a, contents.as_slice())
         .expect("encryption failed");
@@ -97,11 +126,9 @@ fn encrypt_cmd(file_path: &PathBuf, output_path: Option<String>) {
         None => PathBuf::from(default_output_filename),
     };
 
-    // 5. generate a random salt
     let mut salt = [0u8; 16];
     rand::rngs::OsRng.fill_bytes(&mut salt);
 
-    // 6. use the provided password and random salt to generate a derived key
     let mut derived_key = [0u8; 32];
     Argon2::new(
         Argon2d,
@@ -111,63 +138,72 @@ fn encrypt_cmd(file_path: &PathBuf, output_path: Option<String>) {
     .hash_password_into(&password.as_bytes(), &salt, &mut derived_key)
     .expect("Key derivation failed");
 
-    // 7. encrypt the og_key using the password-derived key
     let cipher_b = XChaCha20Poly1305::new((&derived_key).into());
     let nonce_b = XChaCha20Poly1305::generate_nonce(&mut OsRng);
 
     let encrypted_key = cipher_b
         .encrypt(&nonce_b, og_key.as_ref())
         .expect("encryption failed");
-
+ 
+    let mut hash_salt = [0u8; 16];
+    rand::rngs::OsRng.fill_bytes(&mut hash_salt);
     let mut hash_bytes = [0u8; 32];
     Argon2::new(
         Argon2d,
         V0x13,
         Params::new(42_699u32, 2u32, 8u32, Some(32)).unwrap(),
     )
-    .hash_password_into(&ciphertext, &salt, &mut hash_bytes)
+    .hash_password_into(&ciphertext, &hash_salt, &mut hash_bytes)
     .expect("Hashing ciphertext failed");
 
     println!("hash of encrypted content: {:?}", hash_bytes);
 
-    // 8. use serde to serialize an instance of EncryptedData
     let encrypted_key = EncryptedData {
         key: encrypted_key,
         hash: hash_bytes.to_vec(),
         nonce_a: nonce_a.to_vec(),
         nonce_b: nonce_b.to_vec(),
         salt: salt.to_vec(),
+        hash_salt: hash_salt.to_vec(),
         filename: file_path.to_string_lossy().to_string(),
+        output_filename: output_path.to_string_lossy().to_string(),
     };
 
     write_metadata(encrypted_key);
 
-    // 9. write the encrypted contents to the output file
-    fs::write(&output_path, &ciphertext).unwrap();
-    println!(
-        "output encrypted data to: {}",
-        &output_path.to_string_lossy()
-    );
-
-    // 10. buncha logging
-    // println!("og key generated: {:?}", og_key);
-    // println!("nonce a generated: {:?}", nonce_a);
-
-    // println!(
-    //     "ciphertext from file contents using og_key and nonce: {:?}",
-    //     ciphertext
-    // );
-
-    // println!("salt generated: {:?}", salt);
-    // println!("password provided: {:?}", password);
-    // println!("derived key: {:?}", derived_key);
-    // println!("nonce b generated: {:?}", nonce_b);
-
-    // println!("encrypted key: {:?}", encrypted_key);
+    if Path::new(&output_path).exists() {
+        println!(
+            "File {} already exists. Do you want to overwrite it with the encrypted contents of {}? (y/n)",
+            output_path.display(),
+            file_path.display(),
+        );
+        let mut input = String::new();
+        io::stdin().read_line(&mut input).unwrap();
+        if input.trim() == "y" {
+            fs::write(&output_path, &ciphertext).unwrap();
+            println!(
+                "output encrypted data to: {}",
+                &output_path.to_string_lossy()
+            );
+            println!("{} was overwritten successfully", output_path.display());
+            return;
+        } else {
+            println!("{} not modified\n\nno encrypted data was output.", file_path.display());
+            println!("If this was a mistake, you would need to run the command again and use `y` to overwrite the file.");
+            return;
+        }
+    } else {
+        fs::write(&output_path, &ciphertext).unwrap();
+        println!(
+            "output encrypted data to: {}",
+            &output_path.to_string_lossy()
+        );
+        println!("decrypted data was written to {} successfully", output_path.display());
+        return;
+    }
 }
 
 fn decrypt_cmd(file_path: &PathBuf, output_path: Option<String>) {
-    // 1. get the contents of the file as an array of bytes
     let contents_vec = fs::read(file_path).unwrap();
 
     // 2. read the qlock_metadata.json file
@@ -175,7 +211,7 @@ fn decrypt_cmd(file_path: &PathBuf, output_path: Option<String>) {
 
     for datum in saved.data {
         let saved_hash = datum.hash;
-        let salt = datum.salt;
+        let hash_salt = datum.hash_salt;
 
         let mut hash_bytes = [0u8; 32];
         Argon2::new(
@@ -183,11 +219,12 @@ fn decrypt_cmd(file_path: &PathBuf, output_path: Option<String>) {
             V0x13,
             Params::new(42_699u32, 2u32, 8u32, Some(32)).unwrap(),
         )
-        .hash_password_into(&contents_vec, &salt, &mut hash_bytes)
+        .hash_password_into(&contents_vec, &hash_salt, &mut hash_bytes)
         .expect("Hashing ciphertext failed");
 
         if hash_bytes == *saved_hash {
             let key = datum.key;
+            let salt = datum.salt;
             let nonce_a = XNonce::from_slice(&datum.nonce_a);
             let nonce_b = XNonce::from_slice(&datum.nonce_b);
             let filename: String = match output_path {
@@ -195,7 +232,6 @@ fn decrypt_cmd(file_path: &PathBuf, output_path: Option<String>) {
                 None => datum.filename,
             };
 
-            // 4. use the salt and password to generate a derived key
             let password = rpassword::prompt_password("Enter password: ").unwrap();
 
             let mut derived_key = [0u8; 32];
@@ -207,24 +243,14 @@ fn decrypt_cmd(file_path: &PathBuf, output_path: Option<String>) {
             .hash_password_into(&password.as_bytes(), &salt, &mut derived_key)
             .expect("Key derivation failed");
 
-            // 5. decrypt the encrypted key using the derived key
             let cipher_b = XChaCha20Poly1305::new((&derived_key).into());
             let decrypted_key = cipher_b.decrypt(nonce_b, &*key).expect("decryption failed");
 
-            // 6. decrypt the contents of the file using the decrypted key
             let cipher_a = XChaCha20Poly1305::new((&*decrypted_key).into());
             let decrypted_contents = cipher_a
                 .decrypt(nonce_a, contents_vec.as_slice())
                 .expect("decryption failed");
 
-            println!(
-                "decrypted contents: {:?}",
-                String::from_utf8_lossy(&decrypted_contents)
-            );
-
-            // 7. write the decrypted contents to a file with the original filename
-            //    - if the file already exists, ask the user if they want to overwrite it
-            //    - otherwise, just write the file and let them know the results.
             if Path::new(&filename).exists() {
                 println!(
                     "File {} already exists. Do you want to overwrite it with the decrypted contents of {}? (y/n)",
@@ -275,7 +301,24 @@ fn write_metadata(additional_metadata: EncryptedData) {
 }
 
 fn read_metadata() -> SavedData {
+    if !Path::new("qlock_metadata.json").exists() {
+        return SavedData { data: vec![] };
+    }
     let metadata = fs::read_to_string("qlock_metadata.json").unwrap();
     let saved_data: SavedData = serde_json::from_str(&metadata).unwrap();
     saved_data
+}
+
+fn list_metadata() {
+    if !Path::new("qlock_metadata.json").exists() {
+        println!("qlock_metadata.json does not exist, try encrypting something first or make sure you're in the correct directory.");
+        return;
+    }
+    let saved_data: SavedData = read_metadata();
+
+    for datum in saved_data.data {
+        println!("hash: {:?}", datum.hash);
+        println!("  filename: {}", datum.filename);
+        println!("  encrypted output: {}", datum.output_filename);
+    }
 }
